+++ mixer/pkg/il/compiled/BUILD
@@ -5,22 +5,16 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "checker.go",
-        "evaluator.go",
+        "expression.go",
+        "expressionbuilder.go",
     ],
     visibility = ["//visibility:public"],
     deps = [
         "//mixer/pkg/attribute:go_default_library",
-        "//mixer/pkg/config:go_default_library",
-        "//mixer/pkg/config/descriptor:go_default_library",
         "//mixer/pkg/expr:go_default_library",
-        "//mixer/pkg/il:go_default_library",
         "//mixer/pkg/il/compiler:go_default_library",
         "//mixer/pkg/il/interpreter:go_default_library",
         "//mixer/pkg/il/runtime:go_default_library",
-        "@com_github_golang_glog//:go_default_library",
-        "@com_github_hashicorp_golang_lru//:go_default_library",
-        "@io_istio_api//mixer/v1/config/descriptor:descriptor",
     ],
 )
 
@@ -28,9 +22,8 @@ go_test(
     name = "go_default_test",
     size = "medium",
     srcs = [
-        "checker_test.go",
-        "evaluatorBenchmark_test.go",
-        "evaluator_test.go",
+        "expressionbuilder_test.go",
+        "expressionBenchmark_test.go",
     ],
     library = ":go_default_library",
     deps = [
