+++ mixer/pkg/il/compiled/expression.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package evaluator
+package compiled
 
 import "istio.io/istio/mixer/pkg/attribute"
